Revert "lazy loading of operational limits group (#93)"
