+++ network-store-server/src/main/java/com/powsybl/network/store/server/Mappings.java
@@ -107,7 +107,6 @@ public class Mappings {
                                                    areaMappings);
 
     private final Map<String, TableMapping> mappingByTable = new LinkedHashMap<>();
-    private final Map<ResourceType, TableMapping> mappingByType = new LinkedHashMap<>();
 
     private static final String VOLTAGE_LEVEL_ID = "voltageLevelId";
     private static final String VOLTAGE_LEVEL_ID_1 = "voltageLevelId1";
@@ -144,15 +143,6 @@ public class Mappings {
         TableMapping tableMapping = mappingByTable.get(table);
         if (tableMapping == null) {
             throw new IllegalArgumentException("Unknown table: " + table);
-        }
-        return tableMapping;
-    }
-
-    public TableMapping getTableMapping(ResourceType type) {
-        Objects.requireNonNull(type);
-        TableMapping tableMapping = mappingByType.get(type);
-        if (tableMapping == null) {
-            throw new IllegalArgumentException("Unknown type: " + type);
         }
         return tableMapping;
     }
@@ -827,7 +817,6 @@ public class Mappings {
         createAreaMappings();
         for (TableMapping tableMapping : all) {
             mappingByTable.put(tableMapping.getTable().toLowerCase(), tableMapping);
-            mappingByType.put(tableMapping.getResourceType(), tableMapping);
         }
     }
 }
