+++ network-store-server/src/main/java/com/powsybl/network/store/server/QueryCatalog.java
@@ -58,8 +58,6 @@ public final class QueryCatalog {
     public static final String SIDE_COLUMN = "side";
     private static final String TYPE_COLUMN = "type";
     static final String REGULATING = "regulating";
-    public static final String SELECTED_OPERATIONAL_LIMITS_GROUP_ID1 = "selectedoperationallimitsgroupid1";
-    public static final String SELECTED_OPERATIONAL_LIMITS_GROUP_ID2 = "selectedoperationallimitsgroupid2";
     private static final Predicate<String> CLONE_PREDICATE = column -> !column.equals(UUID_COLUMN) && !column.equals(VARIANT_ID_COLUMN)
             && !column.equals(NAME_COLUMN) && !column.equals(FULL_VARIANT_NUM_COLUMN);
     private static final String TOMBSTONED_IDENTIFIABLE_TABLE = "tombstonedidentifiable";
@@ -75,15 +73,6 @@ public final class QueryCatalog {
                 " where " + NETWORK_UUID_COLUMN + " = ?" +
                 " and " + VARIANT_NUM_COLUMN + " = ?" +
                 " and " + ID_COLUMN + " = ?";
-    }
-
-    public static String buildGetSelectedOperationalLimitsGroupsQuery(String tableName) {
-        return "select " + ID_COLUMN + ", " +
-            SELECTED_OPERATIONAL_LIMITS_GROUP_ID1 + ", " +
-            SELECTED_OPERATIONAL_LIMITS_GROUP_ID2 +
-            " from " + tableName +
-            " where " + NETWORK_UUID_COLUMN + " = ?" +
-            " and " + VARIANT_NUM_COLUMN + " = ?";
     }
 
     public static String buildGetNetworkQuery(Collection<String> columns) {
@@ -478,7 +467,7 @@ public final class QueryCatalog {
                 " = ? and " + VARIANT_NUM_COLUMN + " = ?";
     }
 
-    public static String buildGetPermanentLimitQuery(String columnNameForWhereClause) {
+    public static String buildPermanentLimitQuery(String columnNameForWhereClause) {
         return "select " + EQUIPMENT_ID_COLUMN + ", " +
                 EQUIPMENT_TYPE_COLUMN + ", " +
                 NETWORK_UUID_COLUMN + ", " +
