+++ network-store-server/src/test/java/com/powsybl/network/store/server/NetworkStoreRepositoryPartialVariantIdentifiablesTest.java
@@ -381,6 +381,10 @@ class NetworkStoreRepositoryPartialVariantIdentifiablesTest {
                 .attributes(LineAttributes.builder()
                         .voltageLevelId1("vl1")
                         .voltageLevelId2("vl3")
+                        .operationalLimitsGroups1(Map.of("group1", OperationalLimitsGroupAttributes.builder()
+                                .id("group1")
+                                .currentLimits(LimitsAttributes.builder().permanentLimit(20.).build())
+                                .build()))
                         .build())
                 .build();
         networkStoreRepository.createLines(NETWORK_UUID, List.of(line2));
@@ -392,6 +396,10 @@ class NetworkStoreRepositoryPartialVariantIdentifiablesTest {
                 .attributes(LineAttributes.builder()
                         .voltageLevelId1("vl1")
                         .voltageLevelId2("vl2")
+                        .operationalLimitsGroups1(Map.of("group1", OperationalLimitsGroupAttributes.builder()
+                                .id("group1")
+                                .currentLimits(LimitsAttributes.builder().permanentLimit(30.).build())
+                                .build()))
                         .build())
                 .build();
 
@@ -401,26 +409,9 @@ class NetworkStoreRepositoryPartialVariantIdentifiablesTest {
         assertEquals(Optional.of(expLine1), networkStoreRepository.getLine(NETWORK_UUID, 1, lineId1));
         assertEquals(Optional.of(line2), networkStoreRepository.getLine(NETWORK_UUID, 1, lineId2));
 
-        Resource<LineAttributes> withoutLimitsLine1 = Resource.lineBuilder()
-            .id(lineId1)
-            .variantNum(1)
-            .attributes(LineAttributes.builder()
-                .voltageLevelId1("vl1")
-                .voltageLevelId2("vl2")
-            .build())
-            .build();
-        Resource<LineAttributes> withoutLimitsLine2 = Resource.lineBuilder()
-            .id(lineId2)
-            .variantNum(1)
-            .attributes(LineAttributes.builder()
-                .voltageLevelId1("vl1")
-                .voltageLevelId2("vl3")
-                .build())
-            .build();
+        assertEquals(List.of(expLine1, line2), networkStoreRepository.getVoltageLevelLines(NETWORK_UUID, 1, "vl1"));
 
-        assertEquals(List.of(withoutLimitsLine1, withoutLimitsLine2), networkStoreRepository.getVoltageLevelLines(NETWORK_UUID, 1, "vl1"));
-
-        assertEquals(List.of(withoutLimitsLine1, withoutLimitsLine2), networkStoreRepository.getLines(NETWORK_UUID, 1));
+        assertEquals(List.of(expLine1, line2), networkStoreRepository.getLines(NETWORK_UUID, 1));
     }
 
     @Test
