--- network-store-server/src/test/resources/OneOperationalLimitsGroupResultTest.json
@@ -1,19 +1,0 @@
-{
-  "data": [
-    {
-      "id": "selectedside1",
-      "currentLimits": {
-        "permanentLimit": 1.0,
-        "temporaryLimits": {
-          "10": {
-            "name": "temporarylimit1",
-            "value": 12.0,
-            "acceptableDuration": 10,
-            "fictitious": false
-          }
-        }
-      }
-    }
-  ],
-  "meta": {}
-}
