--- network-store-server/src/test/resources/allOperationalLimitsGroupForBranchLine1ResultTest.json
@@ -1,30 +1,0 @@
-[
-  {
-    "id": "selectedside1",
-    "currentLimits": {
-      "permanentLimit": 1.0,
-      "temporaryLimits": {
-        "10": {
-          "name": "temporarylimit1",
-          "value": 12.0,
-          "acceptableDuration": 10,
-          "fictitious": false
-        }
-      }
-    }
-  },
-  {
-    "id": "otherside1line1",
-    "currentLimits": {
-      "permanentLimit": 1.0,
-      "temporaryLimits": {
-        "10": {
-          "name": "temporarylimit1",
-          "value": 12.0,
-          "acceptableDuration": 10,
-          "fictitious": false
-        }
-      }
-    }
-  }
-]
