--- network-store-server/src/test/resources/allOperationalLimitsGroupResultTest.json
@@ -1,98 +1,0 @@
-{
-  "line2": {
-    "1": {
-      "selectedside1": {
-        "id": "selectedside1",
-        "currentLimits": {
-          "permanentLimit": 1.0,
-          "temporaryLimits": {
-            "10": {
-              "name": "temporarylimit1",
-              "value": 12.0,
-              "acceptableDuration": 10,
-              "fictitious": false
-            }
-          }
-        }
-      }
-    },
-    "2": {
-      "otherside1line2": {
-        "id": "otherside1line2",
-        "currentLimits": {
-          "permanentLimit": 1.0,
-          "temporaryLimits": {
-            "10": {
-              "name": "temporarylimit1",
-              "value": 12.0,
-              "acceptableDuration": 10,
-              "fictitious": false
-            }
-          }
-        }
-      },
-      "selectedside2": {
-        "id": "selectedside2",
-        "currentLimits": {
-          "permanentLimit": 1.0,
-          "temporaryLimits": {
-            "10": {
-              "name": "temporarylimit1",
-              "value": 12.0,
-              "acceptableDuration": 10,
-              "fictitious": false
-            }
-          }
-        }
-      }
-    }
-  },
-  "line1": {
-    "1": {
-      "otherside1line1": {
-        "id": "otherside1line1",
-        "currentLimits": {
-          "permanentLimit": 1.0,
-          "temporaryLimits": {
-            "10": {
-              "name": "temporarylimit1",
-              "value": 12.0,
-              "acceptableDuration": 10,
-              "fictitious": false
-            }
-          }
-        }
-      },
-      "selectedside1": {
-        "id": "selectedside1",
-        "currentLimits": {
-          "permanentLimit": 1.0,
-          "temporaryLimits": {
-            "10": {
-              "name": "temporarylimit1",
-              "value": 12.0,
-              "acceptableDuration": 10,
-              "fictitious": false
-            }
-          }
-        }
-      }
-    },
-    "2": {
-      "selectedside2": {
-        "id": "selectedside2",
-        "currentLimits": {
-          "permanentLimit": 1.0,
-          "temporaryLimits": {
-            "10": {
-              "name": "temporarylimit1",
-              "value": 12.0,
-              "acceptableDuration": 10,
-              "fictitious": false
-            }
-          }
-        }
-      }
-    }
-  }
-}
