+++ pom.xml
@@ -49,7 +49,7 @@
 
     <properties>
         <sirocco.version>1.0</sirocco.version>
-        <powsybl-ws-dependencies.version>2.22.0</powsybl-ws-dependencies.version>
+        <powsybl-ws-dependencies.version>2.21.0</powsybl-ws-dependencies.version>
         <!-- FIXME : to remove when sonar version is updated on github actions -->
         <!-- https://community.sonarsource.com/t/stackoverflowerror-at-defaultinputcomponent-equals/20324 -->
         <!-- The versions are very different from this post. But the fix works again. Maybe a similar problem in sonar code -->
@@ -74,7 +74,6 @@
 
     <dependencyManagement>
         <dependencies>
-            <!-- imports -->
             <dependency>
                 <groupId>com.powsybl</groupId>
                 <artifactId>powsybl-ws-dependencies</artifactId>
